feat: redirect logging to the stdout
